10-6/requirements.txt
# Core dependencies for property scraper application

# Web automation
selenium==4.15.2
beautifulsoup4==4.12.2

# Excel handling
pandas==2.1.3
openpyxl==3.1.2

# Environment configuration
python-dotenv==1.0.0

# HTML parsing
lxml==4.9.3

# Utilities
requests==2.31.0

# Optional: For better performance
# numpy==1.26.2
